+++ package.json
@@ -1,27 +1,23 @@
 {
   "name": "sisteam-2022",
-  "version": "0.0.1",
+  "version": "1.0.1",
+  "private": true,
   "scripts": {
-    "dev": "svelte-kit dev --host",
-    "build": "svelte-kit build",
-    "package": "svelte-kit package",
-    "preview": "svelte-kit preview",
-    "check": "svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch"
+	  "dev": "vite dev",
+	  "build": "vite build",
+	  "preview": "vite preview"
   },
   "devDependencies": {
-    "@sveltejs/adapter-auto": "next",
-    "@sveltejs/kit": "next",
-    "svelte": "^3.46.0",
-    "svelte-check": "^2.2.6",
-    "svelte-preprocess": "^4.10.1"
+		"@sveltejs/adapter-auto": "next",
+		"@sveltejs/kit": "next",
+		"svelte": "^3.44.0",
+		"vite": "^3.1.0-beta.1"
   },
   "type": "module",
   "dependencies": {
-    "@lukeed/uuid": "^2.0.0",
-    "dotenv": "^16.0.0",
-    "firebase": "^9.6.8",
-    "geofire-common": "^5.2.0",
-    "leaflet": "^1.7.1"
+    "dotenv": "^16.0.2",
+    "firebase": "^9.9.4",
+    "geofire-common": "^6.0.0",
+    "leaflet": "^1.8.0"
   }
 }
