+++ src/routes/[user]/+layout.reset.svelte
