--- src/routes/[user]/__error.svelte
@@ -1,16 +1,0 @@
-<script context="module">
-  export function load({ error, status }) {
-      return {
-          props: {
-              title: `${status}: ${error.message}`
-          }
-      };
-  }
-</script>
-
-<script>
-  export let title;
-</script>
-
-
-<h1>{title}</h1>
