+++ src/routes/[user]/estadisticas/+page.svelte
