+++ vite.config.js
@@ -1,0 +1,7 @@
+import { sveltekit } from '@sveltejs/kit/vite';
+
+const config = {
+	plugins: [sveltekit()]
+};
+
+export default config;
